+++ component/azstorage/block_blob_test.go
@@ -280,10 +280,10 @@ func (s *blockBlobTestSuite) TestDefault() {
 	s.assert.EqualValues(AccessTiers["none"], s.az.stConfig.defaultTier)
 	s.assert.EqualValues(0, s.az.stConfig.cancelListForSeconds)
 
-	s.assert.EqualValues(7, s.az.stConfig.maxRetries)
+	s.assert.EqualValues(5, s.az.stConfig.maxRetries)
 	s.assert.EqualValues(900, s.az.stConfig.maxTimeout)
-	s.assert.EqualValues(0, s.az.stConfig.backoffTime)
-	s.assert.EqualValues(0, s.az.stConfig.maxRetryDelay)
+	s.assert.EqualValues(4, s.az.stConfig.backoffTime)
+	s.assert.EqualValues(60, s.az.stConfig.maxRetryDelay)
 
 	s.assert.Empty(s.az.stConfig.proxyAddress)
 }