+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 - [#968](https://github.com/Azure/azure-storage-fuse/issues/968) Duplicate directory listing
 - [#964](https://github.com/Azure/azure-storage-fuse/issues/964) Rename for FNS account failing with source does not exists error
 - [#972](https://github.com/Azure/azure-storage-fuse/issues/972) Mount all fails 
+- Added support for "-o nonempty" to mount on a non-empty mount path
 
 **Breaking Changes**
-- Defaults for retry policy changed. Max retries: 3 to 7, Retry delay 3600 to 900 seconds, Max retry delay 3 to 120 seconds
+- Defaults for retry policy changed. Max retries: 3 to 5, Retry delay: 3600 to 900 seconds, Max retry delay: 4 to 60 seconds
+
+**Features**
+- Added new CLI parameter "--subdirectory=" to mount only a subdirectory from given container 
+
 
 ## 2.0.0-preview.4 (2022-11-03)
 **Breaking Changes**